Backported upgrades from Ark:SE image

+++ scripts/Server-Tools/Config-Tools.ps1
@@ -11,10 +11,13 @@ function Copy-Configs
   foreach ($file in $configFiles.GetEnumerator())
   {
     $sourceFile = (Join-Path $defaultConfigDirectory $file.Name)
-    if (-Not (Test-Path $file.Value)) { mkdir -p $($file.Value) } # New-Item does really weird things in here
-    Copy-Item $sourceFile $file.Value
-  }
-}
+    if (Test-Path $sourceFile)
+    {
+      if (-Not (Test-Path $file.Value)) { mkdir -p $($file.Value) } # New-Item does really weird things in here
+      Copy-Item $sourceFile $file.Value
+    } # if (Test-Path $sourceFile)
+  } # foreach ($file in $configFiles.GetEnumerator())
+} # function Copy-Configs
 
 function Update-ConfigFile
 {
